+++ components/tidb_query/src/aggr_fn/impl_max_min.rs
@@ -4,7 +4,7 @@ use std::cmp::Ordering;
 use std::convert::TryFrom;
 
 use tidb_query_codegen::AggrFunction;
-use tidb_query_datatype::{Collation, EvalType, FieldTypeAccessor};
+use tidb_query_datatype::{Collation, EvalType, FieldTypeAccessor, FieldTypeFlag};
 use tipb::{Expr, ExprType, FieldType};
 
 use crate::codec::collation::*;
@@ -65,6 +65,11 @@ impl<T: Extremum> super::AggrDefinitionParser for AggrFnDefinitionParserExtremum
 
         let out_ft = aggr_def.take_field_type();
         let out_et = box_try!(EvalType::try_from(out_ft.as_accessor().tp()));
+        let is_unsigned = child
+            .get_field_type()
+            .as_accessor()
+            .flag()
+            .contains(FieldTypeFlag::UNSIGNED);
         let out_coll = box_try!(out_ft.as_accessor().collation());
 
         if out_et != eval_type {
@@ -87,6 +92,13 @@ impl<T: Extremum> super::AggrDefinitionParser for AggrFnDefinitionParserExtremum
                 C, match out_coll {
                     Collation::C => Ok(Box::new(AggFnExtremumForBytes::<C, T>::new()))
                 }
+            };
+        }
+
+        if eval_type == EvalType::Int {
+            return match is_unsigned {
+                false => Ok(Box::new(AggFnExtremumForInt::<T, false>::new())),
+                true => Ok(Box::new(AggFnExtremumForInt::<T, true>::new())),
             };
         }
 
@@ -263,6 +275,90 @@ where
     }
 }
 
+#[derive(Debug, AggrFunction)]
+#[aggr_function(state = AggFnStateExtremumForInt::<E, IS_UNSIGNED>::new())]
+pub struct AggFnExtremumForInt<E, const IS_UNSIGNED: bool>
+where
+    E: Extremum,
+    VectorValue: VectorValueExt<Int>,
+{
+    _phantom: std::marker::PhantomData<E>,
+}
+
+impl<E, const IS_UNSIGNED: bool> AggFnExtremumForInt<E, IS_UNSIGNED>
+where
+    E: Extremum,
+    VectorValue: VectorValueExt<Int>,
+{
+    pub fn new() -> Self {
+        Self {
+            _phantom: std::marker::PhantomData,
+        }
+    }
+}
+
+#[derive(Debug)]
+pub struct AggFnStateExtremumForInt<E, const IS_UNSIGNED: bool>
+where
+    E: Extremum,
+    VectorValue: VectorValueExt<Int>,
+{
+    extremum: Option<Int>,
+    _phantom: std::marker::PhantomData<E>,
+}
+
+impl<E, const IS_UNSIGNED: bool> AggFnStateExtremumForInt<E, IS_UNSIGNED>
+where
+    E: Extremum,
+    VectorValue: VectorValueExt<Int>,
+{
+    pub fn new() -> Self {
+        Self {
+            extremum: None,
+            _phantom: std::marker::PhantomData,
+        }
+    }
+}
+
+impl<E, const IS_UNSIGNED: bool> super::ConcreteAggrFunctionState
+    for AggFnStateExtremumForInt<E, IS_UNSIGNED>
+where
+    E: Extremum,
+    VectorValue: VectorValueExt<Int>,
+{
+    type ParameterType = Int;
+
+    #[inline]
+    fn update_concrete(&mut self, _ctx: &mut EvalContext, value: &Option<Int>) -> Result<()> {
+        if value.is_some() {
+            if self.extremum.is_none() {
+                self.extremum = value.clone();
+                return Ok(());
+            }
+            if IS_UNSIGNED {
+                let v1 = self.extremum.map(|x| x as u64);
+                let v2 = value.map(|x| x as u64);
+                if v1.cmp(&v2) == E::ORD {
+                    self.extremum = value.clone()
+                }
+            } else {
+                let v1 = self.extremum.map(|x| x as i64);
+                let v2 = value.map(|x| x as i64);
+                if v1.cmp(&v2) == E::ORD {
+                    self.extremum = value.clone()
+                }
+            }
+        }
+        Ok(())
+    }
+
+    #[inline]
+    fn push_result(&self, _ctx: &mut EvalContext, target: &mut [VectorValue]) -> Result<()> {
+        target[0].push(self.extremum);
+        Ok(())
+    }
+}
+
 #[cfg(test)]
 mod tests {
     use tidb_query_datatype::EvalType;
@@ -272,7 +368,7 @@ mod tests {
     use crate::aggr_fn::parser::AggrDefinitionParser;
     use crate::aggr_fn::AggrFunction;
     use crate::codec::batch::{LazyBatchColumn, LazyBatchColumnVec};
-    use tidb_query_datatype::{FieldTypeAccessor, FieldTypeTp};
+    use tidb_query_datatype::{FieldTypeAccessor, FieldTypeFlag, FieldTypeTp};
 
     #[test]
     fn test_max() {
@@ -427,20 +523,6 @@ mod tests {
 
     #[test]
     fn test_integration() {
-        let max_parser = AggrFnDefinitionParserExtremum::<Max>::new();
-        let min_parser = AggrFnDefinitionParserExtremum::<Min>::new();
-
-        let max = ExprDefBuilder::aggr_func(ExprType::Max, FieldTypeTp::LongLong)
-            .push_child(ExprDefBuilder::column_ref(0, FieldTypeTp::LongLong))
-            .build();
-        max_parser.check_supported(&max).unwrap();
-
-        let min = ExprDefBuilder::aggr_func(ExprType::Min, FieldTypeTp::LongLong)
-            .push_child(ExprDefBuilder::column_ref(0, FieldTypeTp::LongLong))
-            .build();
-        min_parser.check_supported(&min).unwrap();
-
-        let src_schema = [FieldTypeTp::LongLong.into()];
         let mut columns = LazyBatchColumnVec::from(vec![{
             let mut col = LazyBatchColumn::decoded_with_capacity_and_tp(0, EvalType::Int);
             col.mut_decoded().push_int(Some(10000));
@@ -454,6 +536,94 @@ mod tests {
             col
         }]);
         let logical_rows = vec![3, 2, 6, 5, 1, 7];
+        let expected_res = vec![Some(99), Some(-1i64)];
+        let mut field_type = FieldType::default();
+        let fta = field_type.as_mut_accessor();
+        fta.set_tp(FieldTypeTp::LongLong);
+
+        test_integration_util(
+            field_type.tp(),
+            field_type.flag(),
+            &mut columns,
+            &logical_rows,
+            &expected_res,
+        );
+    }
+
+    #[test]
+    fn test_aggr_unsigned_signed_int() {
+        // test unsigned bigint
+        let mut columns = LazyBatchColumnVec::from(vec![{
+            let mut col = LazyBatchColumn::decoded_with_capacity_and_tp(0, EvalType::Int);
+            col.mut_decoded().push_int(None);
+            col.mut_decoded().push_int(Some(-1));
+            col.mut_decoded().push_int(Some(2420174916247255494));
+            col.mut_decoded().push_int(Some(3899490809029152765));
+            col
+        }]);
+        let logical_rows = vec![0, 1, 2, 3];
+        let expected_res = vec![Some(-1), Some(2420174916247255494)];
+        let mut field_type = FieldType::default();
+        let fta = field_type.as_mut_accessor();
+        fta.set_tp(FieldTypeTp::LongLong);
+        fta.set_flag(FieldTypeFlag::UNSIGNED);
+
+        test_integration_util(
+            field_type.tp(),
+            field_type.flag(),
+            &mut columns,
+            &logical_rows,
+            &expected_res,
+        );
+
+        // test signed bigint
+        let mut columns = LazyBatchColumnVec::from(vec![{
+            let mut col = LazyBatchColumn::decoded_with_capacity_and_tp(0, EvalType::Int);
+            col.mut_decoded().push_int(None);
+            col.mut_decoded().push_int(Some(-1));
+            col.mut_decoded().push_int(Some(2420174916247255494));
+            col.mut_decoded().push_int(Some(3899490809029152765));
+            col
+        }]);
+        let logical_rows = vec![0, 1, 2, 3];
+        let expected_res = vec![Some(3899490809029152765), Some(-1)];
+        let mut field_type = FieldType::default();
+        let fta = field_type.as_mut_accessor();
+        fta.set_tp(FieldTypeTp::LongLong);
+
+        test_integration_util(
+            field_type.tp(),
+            field_type.flag(),
+            &mut columns,
+            &logical_rows,
+            &expected_res,
+        );
+    }
+
+    fn test_integration_util(
+        field_type: FieldTypeTp,
+        child_flag: FieldTypeFlag,
+        columns: &mut LazyBatchColumnVec,
+        logical_rows: &[usize],
+        expected_res: &[Option<Int>],
+    ) {
+        let max_parser = AggrFnDefinitionParserExtremum::<Max>::new();
+        let min_parser = AggrFnDefinitionParserExtremum::<Min>::new();
+
+        let mut child_field_type: FieldType = field_type.into();
+        child_field_type.as_mut_accessor().set_flag(child_flag);
+
+        let max = ExprDefBuilder::aggr_func(ExprType::Max, field_type)
+            .push_child(ExprDefBuilder::column_ref(0, child_field_type.clone()))
+            .build();
+        max_parser.check_supported(&max).unwrap();
+
+        let min = ExprDefBuilder::aggr_func(ExprType::Min, field_type)
+            .push_child(ExprDefBuilder::column_ref(0, child_field_type.clone()))
+            .build();
+        min_parser.check_supported(&min).unwrap();
+
+        let src_schema = [child_field_type];
 
         let mut schema = vec![];
         let mut exp = vec![];
@@ -463,14 +633,14 @@ mod tests {
             .parse(max, &mut ctx, &src_schema, &mut schema, &mut exp)
             .unwrap();
         assert_eq!(schema.len(), 1);
-        assert_eq!(schema[0].as_accessor().tp(), FieldTypeTp::LongLong);
+        assert_eq!(schema[0].as_accessor().tp(), field_type);
         assert_eq!(exp.len(), 1);
 
         let min_fn = min_parser
             .parse(min, &mut ctx, &src_schema, &mut schema, &mut exp)
             .unwrap();
         assert_eq!(schema.len(), 2);
-        assert_eq!(schema[1].as_accessor().tp(), FieldTypeTp::LongLong);
+        assert_eq!(schema[1].as_accessor().tp(), field_type);
         assert_eq!(exp.len(), 2);
 
         let mut ctx = EvalContext::default();
@@ -482,7 +652,13 @@ mod tests {
         // max
         {
             let max_result = exp[0]
-                .eval(&mut ctx, &src_schema, &mut columns, &logical_rows, 6)
+                .eval(
+                    &mut ctx,
+                    &src_schema,
+                    columns,
+                    &logical_rows,
+                    logical_rows.len(),
+                )
                 .unwrap();
             let max_result = max_result.vector_value().unwrap();
             let max_slice: &[Option<Int>] = max_result.as_ref().as_ref();
@@ -495,7 +671,13 @@ mod tests {
         // min
         {
             let min_result = exp[0]
-                .eval(&mut ctx, &src_schema, &mut columns, &logical_rows, 6)
+                .eval(
+                    &mut ctx,
+                    &src_schema,
+                    columns,
+                    &logical_rows,
+                    logical_rows.len(),
+                )
                 .unwrap();
             let min_result = min_result.vector_value().unwrap();
             let min_slice: &[Option<Int>] = min_result.as_ref().as_ref();
@@ -505,7 +687,7 @@ mod tests {
             min_state.push_result(&mut ctx, &mut aggr_result).unwrap();
         }
 
-        assert_eq!(aggr_result[0].as_int_slice(), &[Some(99), Some(-1i64),]);
+        assert_eq!(aggr_result[0].as_int_slice(), &(*expected_res));
     }
 
     #[test]
