+++ components/tidb_query/src/lib.rs
@@ -26,6 +26,7 @@
 #![allow(clippy::no_effect)]
 #![feature(box_patterns)]
 #![feature(matches_macro)]
+#![feature(const_generics)]
 
 #[macro_use]
 extern crate failure;
